refactor(#3038): extract run_bot S2 bootstrap helpers
